change database schema

+++ migrations/20220315150438_initialization.ts
@@ -121,7 +121,7 @@ export async function up(knex: Knex): Promise<void> {
             tab.string("thumbnail_url", 256).notNullable();
             tab.string("description", 256).notNullable().defaultTo("");
             tab.integer("image_type_id").defaultTo(null);
-            tab.tinyint("status").notNullable().defaultTo(0);
+            tab.smallint("status").notNullable().defaultTo(0);
 
             tab.foreign("image_type_id")
                 .references("id")
@@ -191,7 +191,7 @@ export async function up(knex: Knex): Promise<void> {
                 tab.integer("of_region_id").notNullable();
                 tab.integer("by_user_id").notNullable().defaultTo(0);
                 tab.bigInteger("operation_time").notNullable();
-                tab.tinyint("operation_type").notNullable();
+                tab.smallint("operation_type").notNullable();
 
                 tab.foreign("of_region_id")
                     .references("id")
@@ -265,7 +265,7 @@ export async function up(knex: Knex): Promise<void> {
             (tab) => {
                 tab.increments("id", { primaryKey: true });
                 tab.integer("of_image_id").notNullable();
-                tab.tinyint("at_status").notNullable();
+                tab.smallint("at_status").notNullable();
                 tab.integer("drawn_by_user_id").notNullable().defaultTo(0);
                 tab.integer("labeled_by_user_id").notNullable().defaultTo(0);
                 tab.binary("border").notNullable();
